backend/requirements.txt
uvicorn
fastapi
prisma
bcrypt
PyJWT